requirements.txt
streamlit==1.28.2
google-generativeai==0.2.2
